Added Settings activity
The Settings activity  allows to:
- change name;
- change surname.

In this given moment this activity isn't able to:
- change and/or display the picture of the user (is it needed?)

The activity, to display the user name and surname uses the strings saved in the strings.xml file (located under the valuse folder). In order to display the names correctly we need to change the strings value when the settings page is loaded.
In order to perform such task we need to coordinate with Nicola and Thomas.