Re-Designed activity_visualizeActivity
Changed the name of VisualizeActivityActivity in activity_visualizeactivities in order to keep consistency with the other file names. (If is an activity the file name is activity_filename, if it is a fragment then fragment_filename)

Modified the activity where the information of a certain activity is shown. I re-designed the screen entirely but the work there is not finished.
I still need to implement the list of users participating to a certain event.