Created Activity_visualizeactivity_visualizeparticipants
This new activity has the purpose to show the list of participants to a specific activity. This activity can be reached clicking on the button "visualize participants" located on every activity_visualize_activity screen.

The list was created using the ListAdapter.java class and the list_item.xml file created by @AlessandroConte.

The activity also offers a back button to head back to the activity_visualize_activity screen once you are finished watching the participants list.