Update activity_visualizeactivities_visualizeparticipants
This update changes the information showed on the participants list of an activity. Now you can visualize only the prifile picture, name and surname and their phone number.

This was achieved by duplicating and modifying the ListAdapter created by Alessandro Conte.

My version is called listadapter_user and can also be used in the search fragment of the application.