Updated settings activity and created the manage children activity
The settings activity was utilizing an xml file with deprecated tags to define the different elements present on that screen. Now, all of the grafical elements are included in the main xml file and almost all of the attributes of a user are viewed (and can also be updated) by the screen. (Only GUI working, no connection to firebase present at the moment).

The activity manage children was created to view all of the children that the parent has. This screen is accessible by the settings activity via a button. This new screen also has a + button to add new children. This new screen will be developed soon.