Prompt to view Guide page
Added an AlertDialog that prompts the user that just logged into the app to view the guide page.
This prompt doesn't happen when the user already logged before.

I obtained this via the use of a java class (JustLoggedIn.java) that saves a Boolean value indicating if the user just logged in or not.
Then this bool is used to create (or to not create) the AlertDialog in the MainActivity.java .